Fixed: download count issue when product information is not available from solr(#617)
When downloading closed count report and if for some product the information is not available from solr, then in that case we get undefined in the item details resulting in jsonToCsv parse failure.
Removed the check for productId from solr when preparing the download and added the check on specific fields value.
If the product information is not available from solr then the downloaded report does not contain any product specific data for that product, so in that case by default added productId from the count